1. 安裝npm; 2. 安裝 Laravel 的 ui 工具套件： 3. 建立含有使用者認證的bootstrap框架 4. 安裝相依性套件並編譯 5. 編輯 app/User.php，加入三種帳號角色名稱的常數字串，並將 role 這個欄位加入 fillable 中。 6. 編輯 database/migrations/2014_10_12_000000_create_users_table.php 這一個建立使用者資料表的 migration 設定檔，在其中加入一欄用於儲存帳號角色的 role 欄位。 7. 編輯 app/Http/Controllers/Auth/RegisterController.php，設定使用者註冊時的角色都是一般使用者。 8. 建立 Gate 權限規則。 9.生成三個 users: admin, manager, and user ==========================以下開始根據個帳號權限進行設定========================= user => 顯示 manager => 顯示,修改 admin => 新增,顯示, 修改,刪除
10.更新兩個views： players.index 和teams.index，讓他們針對admin 與 manager 呈現受控管的超連結(新增、修改、刪除)
11.讓每一個路由都必須經過 auth 的 middleware (即必須登入帳號才可以使用)
12.增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用
13.不允許編輯表單給 user 權限使用，但允許 admin或manager使用

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\BrandsController;
 use App\Http\Controllers\HeadphonesController;
+use Illuminate\Support\Facades\Auth;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -13,6 +14,7 @@ use App\Http\Controllers\HeadphonesController;
 | contains the "web" middleware group. Now create something great!
 |
 */
+Route::middleware(['auth'])->group(function () {
 Route::get('/', function () {
     return redirect('brands');
 });
@@ -26,15 +28,15 @@ Route::get('brands/{id}', [BrandsController::class, 'show'])->where('id', '[0-9]
 // 修改單一廠牌表單
 Route::get('brands/{id}/edit', [BrandsController::class, 'edit'])->where('id', '[0-9]+')->name('brands.edit');
 // 刪除單一廠商及旗下耳機資料
-Route::delete('brands/delete/{id}', [BrandsController::class, 'destroy'])->where('id', '[0-9]+')->name('brands.destroy');
-// 新增廠商表單
-Route::get('brands/create', [BrandsController::class, 'create'])->name('brands.create');
-// 修改廠商表單
+Route::delete('brands/delete/{id}', [BrandsController::class, 'destroy'])->where('id', '[0-9]+')->name('brands.destroy')->middleware('can:admin');
+// 新增廠牌表單
+Route::get('brands/create', [BrandsController::class, 'create'])->name('brands.create')->middleware('can:admin');
+// 修改廠牌表單
 Route::get('brands/{id}/edit', [BrandsController::class, 'edit'])->where('id', '[0-9]+')->name('brands.edit');
-// 修改球隊資料
+// 修改廠牌資料
 Route::patch('brands/update/{id}', [BrandsController::class, 'update'])->where('id', '[0-9]+')->name('brands.update');
-// 儲存新球隊資料
-Route::post('brands/store', [BrandsController::class, 'store'])->name('brands.store');
+// 儲存新廠牌資料
+Route::post('brands/store', [BrandsController::class, 'store'])->name('brands.store')->middleware('can:admin');
 
 
 // 顯示所有耳機資料
@@ -50,12 +52,16 @@ Route::get('headphones/{id}', [HeadphonesController::class, 'show'])->where('id'
 // 修改單一耳機表單
 Route::get('headphones/{id}/edit', [HeadphonesController::class, 'edit'])->where('id', '[0-9]+')->name('headphones.edit');
 // 刪除單一耳機資料
-Route::delete('headphones/delete/{id}', [HeadphonesController::class, 'destroy'])->where('id', '[0-9]+')->name('headphones.destroy');
+Route::delete('headphones/delete/{id}', [HeadphonesController::class, 'destroy'])->where('id', '[0-9]+')->name('headphones.destroy')->middleware('can:admin');
 // 新增耳機表單
-Route::get('headphones/create', [HeadphonesController::class, 'create'])->name('headphones.create');
+Route::get('headphones/create', [HeadphonesController::class, 'create'])->name('headphones.create')->middleware('can:admin');
 // 修改耳機表單
 Route::get('headphones/{id}/edit', [HeadphonesController::class, 'edit'])->where('id', '[0-9]+')->name('headphones.edit');
 // 修改耳機資料
 Route::patch('headphones/update/{id}', [HeadphonesController::class, 'update'])->where('id', '[0-9]+')->name('headphones.update');
 // 儲存新耳機資料
-Route::post('headphones/store', [HeadphonesController::class, 'store'])->where('id', '[0-9]+')->name('headphones.store');
+Route::post('headphones/store', [HeadphonesController::class, 'store'])->where('id', '[0-9]+')->name('headphones.store')->middleware('can:admin');
+});
+
+Auth::routes();
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
